+++ lumigator/backend/backend/tests/conftest.py
@@ -8,6 +8,8 @@ from unittest.mock import MagicMock, patch
 from uuid import UUID
 
 import boto3
+import evaluator
+import evaluator.interface
 import fsspec
 import pytest
 import requests_mock
@@ -431,8 +433,8 @@ def create_job_config() -> JobConfig:
 
     conf = JobConfig(
         job_id=uuid.uuid4(),
-        job_type=JobType.EVALUATION,
-        command=settings.EVALUATOR_COMMAND,
+        job_type=evaluator.interface.JOB_INTERFACE.type,
+        command=evaluator.interface.JOB_INTERFACE.command,
         args=conf_args,
     )
