+++ lumigator/backend/backend/tests/unit/services/test_job_service.py
@@ -11,7 +11,7 @@ from lumigator_schemas.tasks import TaskType
 from pydantic import ValidationError
 
 from backend.services.exceptions.job_exceptions import JobValidationError
-from backend.services.jobs import JobDefinitionInference
+from backend.services.jobs import job_settings_map
 from backend.settings import settings
 from backend.tests.conftest import TEST_SEQ2SEQ_MODEL
 
@@ -30,7 +30,7 @@ def test_set_null_inference_job_params(job_record, job_service):
         return_value="s3://bucket/path/to/dataset",
     ):
         dataset_s3_path = job_service._dataset_service.get_dataset_s3_path(request.dataset)
-        job_config = job_service.job_settings[JobType.INFERENCE].generate_config(
+        job_config = job_settings_map[JobType.INFERENCE].generate_config(
             request, request.dataset, dataset_s3_path, job_service.storage_path
         )
         assert job_config.job.max_samples == -1
@@ -51,7 +51,7 @@ def test_set_explicit_inference_job_params(job_record, job_service):
         return_value="s3://bucket/path/to/dataset",
     ):
         dataset_s3_path = job_service._dataset_service.get_dataset_s3_path(request.dataset)
-        job_config = job_service.job_settings[JobType.INFERENCE].generate_config(
+        job_config = job_settings_map[JobType.INFERENCE].generate_config(
             request, request.dataset, dataset_s3_path, job_service.storage_path
         )
         assert job_config.job.max_samples == 10
